Komunikasi IPC di enkripsi

+++ src/include/master/worker_metrics.h
@@ -3,7 +3,7 @@
 
 #include "master/process.h"
 
-double initialize_metrics(const char *label, worker_metrics_t* metrics, worker_type_t wot, int index);
-status_t new_task_metrics(const char *label, master_context_t *master_ctx, worker_type_t wot, int index);
+double initialize_metrics(const char *label, worker_metrics_t* metrics, worker_type_t wot, uint8_t index);
+status_t new_task_metrics(const char *label, master_context_t *master_ctx, worker_type_t wot, uint8_t index);
 
 #endif
